Force using Java 8 - 11
For me running Java 17 with this results in a crash. I am not sure why.
I did not check 12, 13, 14, 15, or 16 since they are not LTS releases.

See [here][1] for info on the syntax for the version range. This will
accept not only 11.0.x but also 11.x (not that they ever put anything
besides 0 there nowadays).

[1]: https://github.com/tofi86/universalJavaApplicationStub/issues/51#issuecomment-364806551